single nir2lava conversion module (first version)
